Move logic to patch_helper
